Çalıştırma için gerekli değişiklikler

+++ ContosoUniversity/Models/WeatherContext.cs
@@ -58,7 +58,7 @@ namespace ContosoUniversity.Models
     public class Wind
     {
         public float speed { get; set; }
-        public int deg { get; set; }
+        public float deg { get; set; }
     }
 
     public class Clouds
